chore: move `ClassicScript` related code to a new module
